vp: merged top-levels of linux-vp and linux32-vp
The top-levels of linux-vp(rv64) and linux32-vp(rv32) share a lot of
common code. This redundancy makes changes that should apply to both
vps error-prone.

With this we merge these two toplevels to one. The distinction between
linux-vp and linux32-vp is controlled by defines which are set in the
corresponding CMakeLists.

The change also includes a harmonization of the
CombinedMemoryInterface for rv64 to match that of rv32 (mmu optional /
passed by pointer) and according adjustments of tiny64 and tiny64-mc.

--
manfred.schlaegl@jku.at
Institute for Complex Systems, JKU Linz
https://ics.jku.at/